test(validation): split audit-procedure guards into their own module
The audit-procedure guards were appended to test_always_on_corpus_claims.py,
pushing it from 402 to 691 lines. That crossed the taste-lint 500-line error
threshold and raised the whole-repo taste count from 584 to 585, which failed
four separate pre-push jobs.

The two halves guard different documents. Lines 1-402 guard
model-context-doctrine.md; the appended section guards
rule-audit-procedure.md. Splitting on that boundary is cohesive rather than
arbitrary, and it returns the original file to its exact state on main.

The new module imports MIRROR_DIR, REPO_ROOT, _frontmatter, and
measured_always_on from the original rather than redefining them, so the two
guards cannot drift on what always-on means.

Same 39 tests, now 20 plus 19. Taste count back to 584.

Refs #4315

Copilot-Session: 3ab110a4-2e9e-4504-a731-74fa73874896
